+++ hexital/analysis/patterns.py
@@ -30,7 +30,7 @@ def doji(
     if index is None:
         return False
 
-    def _doji_check(indx: int):
+    def _doji(indx: int):
         body = utils.candle_realbody(candles[indx])
 
         high_low_avg = utils.candle_high_low_avg(candles, length, indx)
@@ -40,6 +40,46 @@ def doji(
         return body < 0.1 * high_low_avg
 
     if lookback is None:
-        return _doji_check(index)
+        return _doji(index)
 
-    return any(_doji_check(i) for i in range(len(candles) - lookback, len(candles)))
+    return any(_doji(i) for i in range(len(candles) - lookback, len(candles)))
+
+
+def hammer(
+    candles: List[Candle],
+    length: int = 10,
+    lookback: Optional[int] = None,
+    asint: bool = False,
+    index: Optional[int] = None,
+) -> bool | int:
+    index = validate_index(index, len(candles), -1)
+    if index is None:
+        return False
+
+    def _hammer(indx: int):
+        is_hammer = False
+
+        body = utils.candle_realbody(candles[index])
+        body_middle = candles[index].high - candles[index].low
+
+        body_average = utils.candle_realbody_avg(candles, length, index)
+
+        upper_shadow_avg = utils.candle_shadow_upper_avg(candles, length, index)
+        lower_shadow_avg = utils.candle_shadow_lower_avg(candles, length, index)
+
+        if (
+            body < body_average
+            and utils.candle_shadow_lower(candles[index]) > lower_shadow_avg
+            and utils.candle_shadow_upper(candles[index]) < upper_shadow_avg
+            and abs(body_middle - candles[index - 1].low) < 10
+        ):
+            is_hammer = True
+
+        if asint:
+            return int(is_hammer)
+        return is_hammer
+
+    if lookback is None:
+        return _hammer(index)
+
+    return any(_hammer(i) for i in range(len(candles) - lookback, len(candles)))